+++ registration/settings_example.py
