+++ registration/setup.py
