+++ registration/static/bootstrap/fonts/glyphicons-halflings-regular.ttf
