+++ registration/static/bootstrap/js/bootstrap.js
