+++ registration/static/bootstrap/js/npm.js
