+++ registration/static/img/title.png
