+++ registration/static/index.css
