+++ registration/static/js/index.js
