+++ registration/templates/admin.html
