+++ registration/templates/index.html
@@ -1,0 +1,19 @@
+{% extends "layout.html" %}
+{% block body %}
+  <div class="container">
+
+    <div class="row">
+
+      <div class="col-sm-6 col-sm-offset-3 col-md-4 col-md-offset-4">
+          
+          <button class="mlh-login" onclick="location.href='{{ auth_url }}'">Hackers MLH Login</button>
+
+          <button class="mlh-login" onclick="location.href='{{ url_for('login') }}'">Sponsor/Mentor Login</button>
+
+      </div>
+
+    </div>
+  
+  </div>
+  
+{% endblock %}
