--- registration_upstart.conf
@@ -1,11 +1,0 @@
-description "uWSGI server instance configured to serve registration"
-
-start on runlevel [2345]
-stop on runlevel [!2345]
-
-setuid user
-setgid www-data
-
-env PATH=/opt/registration/bin
-chdir /opt/registration
-exec uwsgi --ini registration.ini
